requirements.txt
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10
python-telegram-bot==20.6
anthropic>=0.8.0
voyageai==0.1.2
chromadb==0.4.18
fastapi==0.103.2
uvicorn==0.23.2
beautifulsoup4==4.12.2
pydantic==2.4.2
schedule==1.2.0 